feat: Add BoidSettingsBar component for controlling boid behavior
This commit adds a new component called BoidSettingsBar, which is responsible for controlling the behavior of the boids in the canvas. It allows the user to toggle the separation rule on and off using a checkbox. The BoidSettingsBar component is integrated into the CanvasComponent and provides a visual representation of the separation rule status.